requirements.txt
streamlit==1.35.0
pillow==10.2.0
opencv-python-headless==4.10.0.82
ultralytics==8.2.29
